requirements.txt
boto3 botocore
mlflow==2.21.3
scikit-learn==1.4.2
requests>=2.31.0,<3
pandas 
psycopg2-binary
dotenv
ipykernel
seaborn
xgboost
